Update Dependencies, Spelling and more aggressive minimizing of apc.jar
Major updates to kotlin, shadow-jar and clikt.

Addition of proguard for a better minimization, brining the shipped apc.jar under 2MB.
