+++ src/cli/src/main/kotlin/de/rub/mobsec/Main.kt
@@ -1,0 +1,6 @@
+package de.rub.mobsec
+
+import com.github.ajalt.clikt.core.context
+import com.github.ajalt.clikt.core.main
+
+fun main(args: Array<String>) = APC().context { exitProcess = { kotlin.system.exitProcess(it) } }.main(args)
