+++ src/core/build.gradle.kts
@@ -1,15 +1,13 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
 plugins {
     kotlin("jvm")
 }
 
-version = "1.0.1"
+version = "1.0.2"
 
-val junitVersion = "5.9.0"
+val junitVersion: String by project
 
 dependencies {
-    implementation(kotlin("stdlib-jdk8"))
+    implementation(kotlin("stdlib"))
 
     testImplementation(kotlin("test"))
     testImplementation(kotlin("test-junit5"))
@@ -21,7 +19,7 @@ tasks.test {
     useJUnitPlatform()
 }
 
-tasks.withType<KotlinCompile> {
-    kotlinOptions.jvmTarget = "1.8"
-    kotlinOptions.freeCompilerArgs = listOf("-progressive", "-Xsanitize-parentheses")
+kotlin {
+    jvmToolchain(17)
+    compilerOptions.progressiveMode.set(true)
 }
