+++ src/gradle.properties
@@ -1,1 +1,9 @@
 kotlin.code.style=official
+
+# main
+kotlinVersion = 2.2.10
+shadowVersion = 9.1.0
+cliktVersion = 5.0.3
+
+# test
+junitVersion = 5.13.4
