--- src/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = "apc"
-
-include("cli", "core")
